pyproject.toml
[tool.poetry]
name = "suikapy"
version = "0.1.0"
description = ""
authors = ["sturdy-robot <pedrenriquegg@hotmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pygame-ce = "^2.5.0"
pymunk = "^6.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
